Move `Validation` to its own file from `BpValidator`
