+++ app/src/main/java/org/simple/clinic/bp/Validation.kt
@@ -1,0 +1,18 @@
+package org.simple.clinic.bp
+
+// TODO: Rename to "Result".
+sealed class Validation {
+  // TODO: Rename to "Valid".
+  data class Success(
+      val systolic: Int,
+      val diastolic: Int
+  ) : Validation()
+
+  object ErrorSystolicEmpty : Validation()
+  object ErrorDiastolicEmpty : Validation()
+  object ErrorSystolicTooHigh : Validation()
+  object ErrorSystolicTooLow : Validation()
+  object ErrorDiastolicTooHigh : Validation()
+  object ErrorDiastolicTooLow : Validation()
+  object ErrorSystolicLessThanDiastolic : Validation()
+}
