+++ app/src/main/java/org/simple/clinic/bp/entry/BloodPressureEntryEffectHandler.kt
@@ -11,15 +11,15 @@ import org.simple.clinic.ReportAnalyticsEvents
 import org.simple.clinic.bp.BloodPressureMeasurement
 import org.simple.clinic.bp.BloodPressureReading
 import org.simple.clinic.bp.BloodPressureRepository
-import org.simple.clinic.bp.entry.BpValidator.Validation
-import org.simple.clinic.bp.entry.BpValidator.Validation.ErrorDiastolicEmpty
-import org.simple.clinic.bp.entry.BpValidator.Validation.ErrorDiastolicTooHigh
-import org.simple.clinic.bp.entry.BpValidator.Validation.ErrorDiastolicTooLow
-import org.simple.clinic.bp.entry.BpValidator.Validation.ErrorSystolicEmpty
-import org.simple.clinic.bp.entry.BpValidator.Validation.ErrorSystolicLessThanDiastolic
-import org.simple.clinic.bp.entry.BpValidator.Validation.ErrorSystolicTooHigh
-import org.simple.clinic.bp.entry.BpValidator.Validation.ErrorSystolicTooLow
-import org.simple.clinic.bp.entry.BpValidator.Validation.Success
+import org.simple.clinic.bp.Validation
+import org.simple.clinic.bp.Validation.ErrorDiastolicEmpty
+import org.simple.clinic.bp.Validation.ErrorDiastolicTooHigh
+import org.simple.clinic.bp.Validation.ErrorDiastolicTooLow
+import org.simple.clinic.bp.Validation.ErrorSystolicEmpty
+import org.simple.clinic.bp.Validation.ErrorSystolicLessThanDiastolic
+import org.simple.clinic.bp.Validation.ErrorSystolicTooHigh
+import org.simple.clinic.bp.Validation.ErrorSystolicTooLow
+import org.simple.clinic.bp.Validation.Success
 import org.simple.clinic.bp.entry.PrefillDate.PrefillSpecificDate
 import org.simple.clinic.facility.Facility
 import org.simple.clinic.facility.FacilityRepository
