+++ app/src/main/java/org/simple/clinic/bp/entry/BpValidator.kt
@@ -1,31 +1,15 @@
 package org.simple.clinic.bp.entry
 
-import org.simple.clinic.bp.entry.BpValidator.Validation.ErrorDiastolicTooHigh
-import org.simple.clinic.bp.entry.BpValidator.Validation.ErrorDiastolicTooLow
-import org.simple.clinic.bp.entry.BpValidator.Validation.ErrorSystolicLessThanDiastolic
-import org.simple.clinic.bp.entry.BpValidator.Validation.ErrorSystolicTooHigh
-import org.simple.clinic.bp.entry.BpValidator.Validation.ErrorSystolicTooLow
-import org.simple.clinic.bp.entry.BpValidator.Validation.Success
+import org.simple.clinic.bp.Validation
+import org.simple.clinic.bp.Validation.ErrorDiastolicTooHigh
+import org.simple.clinic.bp.Validation.ErrorDiastolicTooLow
+import org.simple.clinic.bp.Validation.ErrorSystolicLessThanDiastolic
+import org.simple.clinic.bp.Validation.ErrorSystolicTooHigh
+import org.simple.clinic.bp.Validation.ErrorSystolicTooLow
+import org.simple.clinic.bp.Validation.Success
 import javax.inject.Inject
 
 class BpValidator @Inject constructor() {
-
-  // TODO: Rename to "Result".
-  sealed class Validation {
-    // TODO: Rename to "Valid".
-    data class Success(
-        val systolic: Int,
-        val diastolic: Int
-    ) : Validation()
-
-    object ErrorSystolicEmpty : Validation()
-    object ErrorDiastolicEmpty : Validation()
-    object ErrorSystolicTooHigh : Validation()
-    object ErrorSystolicTooLow : Validation()
-    object ErrorDiastolicTooHigh : Validation()
-    object ErrorDiastolicTooLow : Validation()
-    object ErrorSystolicLessThanDiastolic : Validation()
-  }
 
   fun validate(systolic: String, diastolic: String): Validation {
     val systolicNumber = systolic.trim().toInt()
